2518 moved managing of ingest data source and ingest kick off to progress panel
